chore(team): introduce workspace context
